wip: Add tracing and valuable support, update dependencies
This commit adds support for the tracing and valuable crates, updates
various dependencies, and improves logging and error handling. It also
includes minor code refactoring and the addition of new instrument
attributes for better tracing.

Most importantly this commit fixes the issue with the page not going to
the next page :) (though this needs a fix in the mastodon crate as well)

TODO refactor and properly split up this commit.

https://asciinema.org/a/ClkjYYRbjEhNjS8HwcGeLOS5T